+++ src/utils/format_token.test.ts
@@ -108,6 +108,11 @@ describe('utils: formatNumber', () => {
     expect(results).toBe('123,456,789');
   });
 
+  it('format correctly #9', async () => {
+    const results = formatNumber('123456789.00', 2);
+    expect(results).toBe('123,456,789');
+  });
+
   afterEach(() => {
     jest.clearAllMocks();
   });